+++ src/planning/StealResourcesTask.java
@@ -48,7 +48,7 @@ public class StealResourcesTask extends Task {
 	@Override
 	public double getFlavorMatch(Faction faction) 
 	{
-		int resourceAdjustCap = 2000;
+		double resourceAdjustCap = 2000;
 		double selfAdjustFactor = (1 - (Math.min(faction.getNumResources(), resourceAdjustCap) / resourceAdjustCap)) * 0.5;
 		double enemyAdjustFactor = (Math.min(target.getNumResources(), resourceAdjustCap) / resourceAdjustCap) * 0.5;
 		return ((enemyAdjustFactor + selfAdjustFactor) * ((faction.getAggression() * 2 + faction.getDiplomacy()) / 3.0));
